chore(QYOG-26): pagination 오타 수정
