--- src/containers/Board/PagenationBoard/index.ts
@@ -1,1 +1,0 @@
-export { default as PagenationBoard } from "./PagenationBoard";
