+++ src/containers/Board/PaginationBoard/index.ts
@@ -1,0 +1,1 @@
+export { default as PaginationBoard } from "./PaginationBoard";
