+++ src/containers/Board/index.tsx
@@ -5,4 +5,5 @@
  */
 
 export * from "./InfinityScrollBoard";
-export * from "./PagenationBoard";
+export * from "./PaginationBoard";
+export * from "./DetailBoard";
