--- src/pages/board/pagenation/index.tsx
@@ -1,11 +1,0 @@
-/*
- * Created on Wed Nov 15 2023
- *
- * Copyright (c) 2023 Your Company
- */
-
-import { PagenationBoard } from "@/containers/Board";
-
-export default function PagenationTestPage() {
-  return <PagenationBoard />;
-}
