+++ src/pages/board/pagination/index.tsx
@@ -1,0 +1,11 @@
+/*
+ * Created on Wed Nov 15 2023
+ *
+ * Copyright (c) 2023 Your Company
+ */
+
+import { PaginationBoard } from "@/containers/Board";
+
+export default function PaginationTestPage() {
+  return <PaginationBoard />;
+}
